Simplify GTK3 screenshot logic
This moves the GTK-agnostic screenshot methods to the abstract
base-class so that it can be re-used by the upcoming GTK4 port.

Additionally, the callback consumer is no longer passed to the
`getImageSurface()` method and is instead call directly within the
`traverse` method.
